Redesign Workload and introduce Benchmark trait
