+++ synthesizer/benches/benchmarks/credits/transfer_public_to_private.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{BenchmarkOperations, Operation, SetupOperations, Workload};
+use crate::{Operation, Benchmark};
 
 use console::{
     network::Network,
@@ -25,62 +25,64 @@ use snarkvm_utilities::TestRng;
 
 use std::{marker::PhantomData, str::FromStr};
 
-pub struct MintPublic<N: Network> {
+pub struct TransferPublicToPrivate<N: Network> {
     num_executions: usize,
     phantom: PhantomData<N>,
 }
 
-impl<N: Network> MintPublic<N> {
+impl<N: Network> TransferPublicToPrivate<N> {
     pub fn new(num_executions: usize) -> Self {
         Self { num_executions, phantom: Default::default() }
     }
 }
 
-impl<N: Network> Workload<N> for MintPublic<N> {
+impl<N: Network> Benchmark<N> for TransferPublicToPrivate<N> {
     fn name(&self) -> String {
-        format!("mint_public/{}_executions", self.num_executions)
+        format!("transfer_public_to_private/{}_executions", self.num_executions)
     }
 
-    fn init(&mut self) -> (SetupOperations<N>, BenchmarkOperations<N>) {
+    fn setup_operations(&mut self) -> Vec<Vec<Operation<N>>> {
         // Construct the program.
         let program = Program::from_str(&format!(
             r"
-program mint_public_{}.aleo;
+program transfer_public_to_private_{}.aleo;
 mapping account:
     key left as address.public;
     value right as u64.public;
-function mint_public:
+function transfer_public_to_private:
     input r0 as address.public;
     input r1 as u64.public;
     finalize r0 r1;
-finalize mint_public:
+finalize transfer_public_to_private:
     input r0 as address.public;
     input r1 as u64.public;
     get.or_init account[r0] 0u64 into r2;
-    add r2 r1 into r3;
+    sub r2 r1 into r3;
     set r3 into account[r0];
 ",
             self.num_executions
         ))
-        .unwrap();
-        let setups = vec![vec![Operation::Deploy(Box::new(program))]];
+            .unwrap();
+        vec![vec![Operation::Deploy(Box::new(program))]]
+    }
 
+    fn benchmark_operations(&mut self) -> Vec<Operation<N>> {
         // Initialize storage for the benchmark operations.
         let mut benchmarks = Vec::with_capacity(self.num_executions);
         // Initialize an RNG for generating the operations.
         let rng = &mut TestRng::default();
         // Construct the operations.
         for _ in 0..self.num_executions {
+            let sender = Address::rand(rng);
             benchmarks.push(Operation::Execute(
-                format!("mint_public_{}.aleo", self.num_executions),
-                "mint_public".to_string(),
+                format!("transfer_public_to_private_{}.aleo", self.num_executions),
+                "transfer_public_to_private".to_string(),
                 vec![
-                    Value::Plaintext(Plaintext::from(Literal::Address(Address::rand(rng)))),
+                    Value::Plaintext(Plaintext::from(Literal::Address(sender))),
                     Value::Plaintext(Plaintext::from(Literal::U64(U64::zero()))),
                 ],
             ));
         }
-
-        (setups, benchmarks)
+        benchmarks
     }
 }
