+++ synthesizer/benches/benchmarks/get_or_init.rs
@@ -14,14 +14,15 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{BenchmarkOperations, Operation, SetupOperations, Workload};
+use crate::{Operation, Workload};
 
 use console::network::Network;
 use snarkvm_synthesizer::Program;
 
 use std::{marker::PhantomData, str::FromStr};
+use crate::utilities::Benchmark;
 
-pub struct StaticSet<N: Network> {
+pub struct StaticGetOrInit<N: Network> {
     num_mappings: usize,
     num_commands: usize,
     num_executions: usize,
@@ -29,61 +30,64 @@ pub struct StaticSet<N: Network> {
     phantom: PhantomData<N>,
 }
 
-impl<N: Network> StaticSet<N> {
+impl<N: Network> StaticGetOrInit<N> {
     pub fn new(num_mappings: usize, num_commands: usize, num_executions: usize, num_programs: usize) -> Self {
         Self { num_mappings, num_commands, num_executions, num_programs, phantom: Default::default() }
     }
 }
 
-impl<N: Network> Workload<N> for StaticSet<N> {
+impl<N: Network> Benchmark<N> for StaticGetOrInit<N> {
     fn name(&self) -> String {
         format!(
-            "static_set/{}_mappings/{}_commands/{}_executions/{}_programs",
+            "static_get_or_init/{}_mappings/{}_commands/{}_executions/{}_programs",
             self.num_mappings, self.num_commands, self.num_executions, self.num_programs
         )
     }
 
-    fn init(&mut self) -> (SetupOperations<N>, BenchmarkOperations<N>) {
+    fn setup_operations(&mut self) -> Vec<Vec<Operation<N>>> {
         // Initialize storage for the setup operations.
         let mut operations = Vec::with_capacity(self.num_programs);
         // Construct the operations.
         for i in 0..self.num_programs {
             // Initialize the program string.
-            let mut program_string =
-                format!("program set_{}_{}_{}_{i}.aleo;", self.num_mappings, self.num_commands, self.num_executions);
+            let mut program_string = format!(
+                "program get_or_init_{}_{}_{}_{i}.aleo;",
+                self.num_mappings, self.num_commands, self.num_executions
+            );
             // Add the mappings.
             for j in 0..self.num_mappings {
                 program_string
                     .push_str(&format!("mapping map_{j}:key left as field.public;value right as field.public;"));
             }
-            // Construct the setter function.
-            let mut setter = "function setter:finalize;finalize setter:".to_string();
+            // Construct the init function.
+            let mut init_function = "function init:finalize;finalize init:".to_string();
             // Add the commands.
             for j in 0..self.num_mappings {
                 for k in 0..self.num_commands {
-                    setter.push_str(&format!("set {k}field into map_{j}[{k}field];"));
+                    init_function.push_str(&format!("get.or_init map_{j}[{k}field] {k}field into r{k};"));
                 }
             }
             // Add the functions to the program string.
-            program_string.push_str(&setter);
+            program_string.push_str(&init_function);
             // Construct and add the setup operation.
             operations.push(Operation::Deploy(Box::new(Program::from_str(&program_string).unwrap())));
         }
-        let setups = vec![operations];
+        vec![operations]
+    }
 
+    fn benchmark_operations(&mut self) -> Vec<Operation<N>> {
         // Initialize storage for the benchmark operations.
         let mut benchmarks = Vec::with_capacity(self.num_programs * self.num_executions);
         // Construct the operations.
         for i in 0..self.num_programs {
             for _ in 0..self.num_executions {
                 benchmarks.push(Operation::Execute(
-                    format!("set_{}_{}_{}_{i}.aleo", self.num_mappings, self.num_commands, self.num_executions),
-                    "setter".to_string(),
+                    format!("get_or_init_{}_{}_{}_{i}.aleo", self.num_mappings, self.num_commands, self.num_executions),
+                    "init".to_string(),
                     vec![],
                 ));
             }
         }
-
-        (setups, benchmarks)
+        benchmarks
     }
 }
