+++ synthesizer/benches/utilities/mod.rs
@@ -14,6 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
+pub mod benchmark;
+pub use benchmark::*;
+
 pub mod object_store;
 pub use object_store::*;
